refactor(client): extract table logic
