+++ client/src/components/QualityManagement/Parts.jsx
@@ -1,106 +1,29 @@
 import { DeleteIcon } from '@chakra-ui/icons';
-import {
-  Box,
-  Checkbox,
-  Heading,
-  IconButton,
-  Table,
-  Tag,
-  Tbody,
-  Thead,
-  Tr,
-  useToast,
-} from '@chakra-ui/react';
+import { Box, Checkbox, Heading, IconButton, Table, Tag, Tbody, Thead, Tr } from '@chakra-ui/react';
 import { Paper, TableContainer, TablePagination } from '@material-ui/core';
 import { NoResultImage } from 'components/common/Image.jsx';
 import Loader from 'components/common/Loader';
 import { StyledTableCell, StyledTableHeader, StyledTableRow } from 'components/common/Table.jsx';
 import AddDefectModal from 'components/QualityManagement/AddDefectModal.jsx';
-import React, { Fragment, useState } from 'react';
-import { useQuery } from 'react-query';
-import { deleteDefects, getDefects } from 'utils/api/defect.js';
+import useQualityPartsTable from 'hooks/useQualityPartsTable.jsx';
+import React, { Fragment } from 'react';
 
 const Parts = () => {
-  const { isLoading, isSuccess, data, refetch } = useQuery('defects', getDefects);
-  const [page, setPage] = useState(0);
-  const [rowsPerPage, setRowsPerPage] = useState(10);
-  const [selected, setSelected] = useState([]);
-  const toast = useToast();
-
-  const handleChangePage = (e, newPage) => {
-    setPage(newPage);
-  };
-
-  const handleChangeRowsPerPage = (e) => {
-    setRowsPerPage(parseInt(e.target.value, 10));
-    setPage(0);
-  };
-
-  const handleStatusColor = (status) => {
-    switch (status) {
-      case 'Solved':
-        return 'green';
-      case 'Pending':
-        return 'orange';
-      case 'Ongoing':
-        return 'yellow';
-    }
-  };
-
-  const isSelected = (_id) => selected.indexOf(_id) !== -1;
-
-  const handleSelectAllClick = (e) => {
-    if (e.target.checked) {
-      const newSelecteds = data.data.map((n) => n._id);
-      setSelected(newSelecteds);
-      return;
-    }
-    setSelected([]);
-  };
-
-  const handleOnChange = (e, _id) => {
-    const selectedIndex = selected.indexOf(_id);
-    let newSelected = [];
-
-    if (selectedIndex === -1) {
-      newSelected = newSelected.concat(selected, _id);
-    } else if (selectedIndex === 0) {
-      newSelected = newSelected.concat(selected.slice(1));
-    } else if (selectedIndex === selected.length - 1) {
-      newSelected = newSelected.concat(selected.slice(0, -1));
-    } else if (selectedIndex > 0) {
-      newSelected = newSelected.concat(
-        selected.slice(0, selectedIndex),
-        selected.slice(selectedIndex + 1)
-      );
-    }
-
-    setSelected(newSelected);
-  };
-
-  const handleDelete = async (e) => {
-    e.preventDefault();
-    try {
-      await deleteDefects(selected);
-      toast({
-        title: 'Deleted',
-        description: 'Defects have been deleted successfully',
-        status: 'success',
-        duration: 9000,
-        isClosable: true,
-      });
-      setSelected([]);
-      refetch();
-    } catch (err) {
-      toast({
-        title: 'Error',
-        description: err.message,
-        status: 'error',
-        duration: 9000,
-        isClosable: true,
-      });
-    }
-  };
+  const {
+    handleChangePage,
+    handleChangeRowsPerPage,
+    handleStatusColor,
+    isSelected,
+    handleSelectAllClick,
+    handleOnChange,
+    handleDelete,
+    isLoading,
+    isSuccess,
+    page,
+    rowsPerPage,
+    data,
+    selected,
+  } = useQualityPartsTable();
 
   if (isLoading) {
     return <Loader />;
@@ -120,6 +43,7 @@ const Parts = () => {
 
   return (
     <Box overflowX="auto">
+      <AddDefectModal />
       <IconButton
         colorScheme="blue"
         variant="outline"
@@ -200,7 +124,6 @@ const Parts = () => {
           onChangeRowsPerPage={handleChangeRowsPerPage}
         />
       </TableContainer>
-      <AddDefectModal />
     </Box>
   );
 };
